# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "openkoutsi API",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/integrations/available": {
      "get": {
        "tags": [
          "integrations"
        ],
        "summary": "Available",
        "description": "Return the list of provider names that have credentials configured on the server.",
        "operationId": "available_api_integrations_available_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/integrations/status": {
      "get": {
        "tags": [
          "integrations"
        ],
        "summary": "Status",
        "description": "Return the list of provider names the current user has connected.",
        "operationId": "status_api_integrations_status_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/integrations/{provider}/connect": {
      "get": {
        "tags": [
          "integrations"
        ],
        "summary": "Connect",
        "description": "Return the OAuth authorization URL for the given provider.",
        "operationId": "connect_api_integrations__provider__connect_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "provider",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Provider"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/integrations/{provider}/callback": {
      "get": {
        "tags": [
          "integrations"
        ],
        "summary": "Callback",
        "description": "Exchange OAuth code for tokens and persist the connection.\n\nStrava / Wahoo redirect the user's browser here, so Bearer auth is not\navailable. We identify the user via the signed JWT in the ``state`` param.",
        "operationId": "callback_api_integrations__provider__callback_get",
        "parameters": [
          {
            "name": "provider",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Provider"
            }
          },
          {
            "name": "code",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "State"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/integrations/{provider}/sync": {
      "post": {
        "tags": [
          "integrations"
        ],
        "summary": "Sync",
        "description": "Trigger a full history import from the given provider in the background.\n\nImports the full history into the user's own DB.",
        "operationId": "sync_api_integrations__provider__sync_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "provider",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Provider"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/integrations/{provider}/sync-zones": {
      "post": {
        "tags": [
          "integrations"
        ],
        "summary": "Sync Zones",
        "description": "Fetch training zones (HR, power) and FTP from the provider and save to the athlete profile.",
        "operationId": "sync_zones_api_integrations__provider__sync_zones_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "provider",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Provider"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/integrations/{provider}/disconnect": {
      "delete": {
        "tags": [
          "integrations"
        ],
        "summary": "Disconnect",
        "description": "Revoke the provider token and remove the stored connection.\n\nPass ``delete_data=true`` (as a query parameter or in the request body) to\nalso permanently delete every activity imported from this provider from the\nuser's data. If that deletion fails, the connection is left in place and the\nrequest fails with ``500`` \u2014 the caller is never told the data is gone unless\nit actually was.",
        "operationId": "disconnect_api_integrations__provider__disconnect_delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "provider",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Provider"
            }
          },
          {
            "name": "delete_data",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Also delete all activities imported from this provider",
              "default": false,
              "title": "Delete Data"
            },
            "description": "Also delete all activities imported from this provider"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/DisconnectRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DisconnectRequest": {
        "properties": {
          "delete_data": {
            "type": "boolean",
            "title": "Delete Data",
            "default": false
          }
        },
        "type": "object",
        "title": "DisconnectRequest",
        "description": "Optional request body for the disconnect endpoint.\n\n``delete_data`` may be supplied either as a query parameter or in this body;\naccepting both means a query-vs-body mismatch from the caller can't silently\ndefault to ``False`` and skip the deletion."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/login"
          }
        }
      }
    }
  }
}